+++ src/plone/app/fhirfield/helpers.py
@@ -38,11 +38,11 @@ def search_fhir_model(model_name, cache=True):
             model_name)
 
     # Trying to get from entire modules
-    from fhirclient import models
+    from fhir import resources
     for importer, modname, ispkg in \
-            pkgutil.walk_packages(models.__path__, models.__name__ + '.',
+            pkgutil.walk_packages(resources.__path__, resources.__name__ + '.',
                                   onerror=lambda x: None):
-        if ispkg or modname.endswith('_tests'):
+        if ispkg or '.tests.' in modname:
             continue
 
         mod = import_module(modname)