+++ src/plone/app/fhirfield/patch.py
@@ -6,8 +6,8 @@ __author__ = 'Md Nazrul Islam <email2nazrul@gmail.com>'
 def monkey_patch_fhir_base_model():
     """" """
     from zope.interface import implementer
-    import fhirclient.models.resource
+    import fhir.resources.resource
     from plone.app.fhirfield.interfaces import IFhirResourceModel
     # We forcely implement IFhirResourceModel
-    fhirclient.models.resource.Resource = \
-        implementer(IFhirResourceModel)(fhirclient.models.resource.Resource)
+    fhir.resources.resource.Resource = \
+        implementer(IFhirResourceModel)(fhir.resources.resource.Resource)